# mes-root/docs/mes-auto_swagger.yaml
openapi: 3.0.0
...
servers:
  - description: DEV
    url: https://localhost:9998
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/hengyi/mes-auto/1.0.0
info:
  description: auto api
  version: "1.0.0"
  title: Mes Auto
  contact:
    email: ixtf1984@gmail.com
...
paths:
...
  /api/grades:
    post:
      tags:
        - admin
      summary: 新建等级
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - config
      summary: 等级列表
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grade'
        '400':
          $ref: '#/components/responses/JException'
  /api/grades/{id}:
    put:
      tags:
        - admin
      summary: 更新等级
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - config
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          $ref: '#/components/responses/JException'
  /api/batches:
    post:
      tags:
        - admin
      summary: 新建批号
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - config
      summary: 批号列表
      parameters:
        - in: query
          name: first
          schema:
            type: integer
            minimum: 0
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 10
            maximum: 100
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  batches:
                    type: array
                    items:
                      $ref: '#/components/schemas/Batch'
        '400':
          $ref: '#/components/responses/JException'
  /api/batches/{id}:
    put:
      tags:
        - admin
      summary: 更新批号
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - config
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          $ref: '#/components/responses/JException'
  /api/products:
    post:
      tags:
        - admin
      summary: 新建产品
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - config
      summary: 产品列表
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/JException'
  /api/products/{id}:
    put:
      tags:
        - admin
      summary: 更新产品
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/JException'
  /api/products/{id}/dyeingInfo:
    put:
      tags:
        - admin
      summary: 更新产品
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDyeingInfoUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - config
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/JException'
...
components:
  securitySchemes:
    # arbitrary name for the security scheme
    bearerAuth:
      type: http
      scheme: bearer
      # optional, arbitrary value for documentation purposes
      bearerFormat: JWT
  responses:
...
    JException:
      description: |
        非200的异常一般都会处理过，没处理的都是http status默认

        如：401没认证，403没权限，404找不到，等等
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JException'
  schemas:
    JException:
      type: object
      properties:
        errorCode:
          type: string
          example: E00000
        errorMessage:
          type: string
...
    EntityDTO:
      type: object
      required:
        - id
      properties:
        id:
          type: string
...
    Corporation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    Workshop:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        note:
          type: string
        corporation:
          $ref: '#/components/schemas/Corporation'
...
    ProductUpdateCommand:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
        code:
          type: string
          minLength: 2
          maxLength: 2
    ProductDyeingInfoUpdateCommand:
      type: object
      properties:
        dyeingFormConfig:
          $ref: '#/components/schemas/EntityDTO'
    Product:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
    BatchUpdateCommand:
      type: object
      required:
        - name
        - sortBy
      properties:
        name:
          type: string
        sortBy:
          type: string
    Batch:
      type: object
      description: 批号
      properties:
        id:
          type: string
        batchNo:
          type: string
        silkWeight:
          type: number
        centralValue:
          type: number
        holeNum:
          type: number
        spec:
          type: string
          description: 规格
        tubeColor:
          type: string
          description: 规格
        note:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        product:
          $ref: '#/components/schemas/Product'
    GradeUpdateCommand:
      type: object
      required:
        - name
        - sortBy
      properties:
        name:
          type: string
        sortBy:
          type: string
    Grade:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sortBy:
          type: string
          description: 分数，丝锭的等级需要100分，才能推入自动包装线
...
security:
  - bearerAuth: []